Remove unused, move conversion method, check exists co2SavingsData
