--- src/app/phast/losses/operations/operations.component.css
@@ -1,3 +1,0 @@
-.modal-open {
-    z-index: initial !important;
-}
